requirements.txt
click==8.1.2
dlib==19.23.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
watchdog==2.1.7
